Refactor download-svg.business.ts and index.ts
