+++ src/common/export/download-svg.business.ts
@@ -1,24 +1,6 @@
 import ReactDOMServer from 'react-dom/server';
 import { Size } from '@/core/model';
-
-const initiateDownload = (url: string, name: string) => {
-  const link = document.createElement('a');
-  link.href = url;
-  link.download = name;
-  document.body.appendChild(link);
-  link.click();
-  document.body.removeChild(link);
-};
-
-export const downloadFile = (
-  filename: string,
-  content: string,
-  contentType: string
-) => {
-  const blob = new Blob([content], { type: contentType });
-  const svgUrl = URL.createObjectURL(blob);
-  initiateDownload(svgUrl, filename);
-};
+import { downloadFile, initiateDownload } from './classic-download';
 
 export const downloadSvg = (svg: JSX.Element) => {
   const svgString = ReactDOMServer.renderToStaticMarkup(svg);
