+++ src/common/export/index.ts
@@ -1,2 +1,3 @@
 export * from './download-svg.business';
 export * from './save-file-modern';
+export * from './classic-download';
